--- compress/BUILD.bazel
@@ -1,14 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "go_default_library",
-    srcs = ["compress.go"],
-    importpath = "github.com/google/go-containerregistry/compress",
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["compress_test.go"],
-    embed = [":go_default_library"],
-)
