+++ v1/tarball/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
     importpath = "github.com/google/go-containerregistry/v1/tarball",
     visibility = ["//visibility:public"],
     deps = [
-        "//compress:go_default_library",
         "//name:go_default_library",
         "//v1:go_default_library",
         "//v1/partial:go_default_library",
